Use the super function to get superclass for ctor
Using super to get the superclass is better than
having to hardcode the superclass into the
ctor directly.

There was a bug, where in the vtable_ctor we were
overwriting the superclass field in a memcpy.
This was fixed, where we simply save the pointer
to the superclass and then write it back in after
the memcpy, but this shows the memcpy is flawed
to begin with and should be fixed.