Bugfix: proper creation of container metaclass
Also, changed the container object to a list
object, which made the semantics a bit easiser
to make sense of. In general, a list won't
derive from container which derives from object,
instead a list derives directly from object, but
it's methods will all derive from container.
(Because a container has begin and end methods).